+++ modules/tinymce/src/themes/silver/main/ts/ui/core/color/ColorSwatch.ts
@@ -112,13 +112,8 @@ const getFetch = (colors: Menu.ChoiceMenuItemSpec[], hasCustom: boolean) => (cal
 };
 
 const setIconColor = (splitButtonApi: Toolbar.ToolbarSplitButtonInstanceApi, name: string, newColor: string) => {
-  const setIconFillAndStroke = (pathId, color) => {
-    splitButtonApi.setIconFill(pathId, color);
-    splitButtonApi.setIconStroke(pathId, color);
-  };
-
   const id = name === 'forecolor' ? 'tox-icon-text-color__color' : 'tox-icon-highlight-bg-color__color';
-  setIconFillAndStroke(id, newColor);
+  splitButtonApi.setIconFill(id, newColor);
 };
 
 const registerTextColorButton = (editor: Editor, name: string, format: string, tooltip: string, lastColor: Cell<string>) => {